+++ lib/onlyoffice_webdriver_wrapper/helpers/headless_helper.rb
@@ -4,10 +4,12 @@ require_relative 'headless_helper/real_display_tools'
 require_relative 'headless_helper/ruby_helper'
 require 'headless'
 require_relative 'headless_helper/headless_screenshot_patch'
+require_relative 'headless_helper/headless_video_recorder'
 
 module OnlyofficeWebdriverWrapper
   # Class for using headless gem
   class HeadlessHelper
+    include HeadlessVideoRecorder
     include RealDisplayTools
     include RubyHelper
     # @return [Headless] instance of headless object
@@ -16,10 +18,15 @@ module OnlyofficeWebdriverWrapper
     attr_accessor :resolution_x
     # @return [Integer] y resolution of virtual screen
     attr_accessor :resolution_y
+    # @return [True, False] is video should be recorded
+    attr_reader :record_video
 
-    def initialize(resolution_x = 1680, resolution_y = 1050)
+    def initialize(resolution_x = 1680,
+                   resolution_y = 1050,
+                   record_video: true)
       @resolution_x = resolution_x
       @resolution_y = resolution_y
+      @record_video = record_video
     end
 
     # Check if should start headless
@@ -43,21 +50,25 @@ module OnlyofficeWebdriverWrapper
       begin
         @headless_instance = Headless.new(reuse: false,
                                           destroy_at_exit: true,
-                                          dimensions: "#{@resolution_x + 1}x#{@resolution_y + 1}x24")
+                                          dimensions: "#{@resolution_x + 1}x#{@resolution_y + 1}x24",
+                                          video: { provider: :ffmpeg })
       rescue Exception => e
         OnlyofficeLoggerHelper.log("xvfb not started with problem #{e}")
         WebDriver.clean_up(true)
         @headless_instance = Headless.new(reuse: false,
                                           destroy_at_exit: true,
-                                          dimensions: "#{@resolution_x + 1}x#{@resolution_y + 1}x24")
+                                          dimensions: "#{@resolution_x + 1}x#{@resolution_y + 1}x24",
+                                          video: { provider: :ffmpeg })
       end
       headless_instance.start
+      start_capture
     end
 
     def stop
       return unless running?
 
       OnlyofficeLoggerHelper.log('Stopping Headless Session')
+      stop_capture
       headless_instance.destroy
     end